+++ include/core/window.h
@@ -4,6 +4,8 @@
 #include "glad/glad.h"
 #include <GLFW/glfw3.h>
 
+#include "core/errorHandler.h"
+
 // default values for window (can be modified later)
 #define WIDTH 800
 #define HEIGHT 600
@@ -16,7 +18,7 @@
 // function declarations
 void framebufferSizeCallback(GLFWwindow *window, int width, int height);
 
-int initGLFW(void);
-int initWindow(GLFWwindow **window);
+enum ErrorCode initGLFW(void);
+enum ErrorCode initWindow(GLFWwindow **window);
 
 #endif // WINDOW_H