+++ include/utils/shaderUtils.h
@@ -2,19 +2,14 @@
 #define SHADER_UTILS_H
 
 #include <GLFW/glfw3.h>
-#include <stdio.h>
 
-FILE *openFile(const char* path);
-long getFileSize(FILE *file);
-char* readFileToString(FILE *file, const long fSize);
+#include "core/errorHandler.h"
 
-// implements all the above functions
-char* loadShaderSource(const char* path);
+enum ErrorCode loadShaderSource(const char *path, char **outSource);
+enum ErrorCode compileShader(const char *source, unsigned int shaderType, GLuint *outShader);
+enum ErrorCode linkShaders(GLuint vertex, GLuint fragment, GLuint *outProgram);
 
-GLuint compileShader(const char* source, unsigned int shaderType);
-GLuint linkShaders(GLuint vertex, GLuint fragment);
-
-void gluSetFloat(GLuint program, const char* name, const float value);
+void gluSetFloat(GLuint program, const char *name, const float value);
 void gluSet4f(GLuint program, const char *name, const float x, const float y, const float z, const float w);
 
 #endif // SHADER_UTILS_H